refactor: SignUpResponse 클래스의 패키지 경로 변경 및 Builder 패턴 제거
